requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.11.0.86